Use a "Sample ID" file to drive data-loading.
The file should contain Sample IDs for a single tissue type.
Go to https://amp.pharm.mssm.edu/archs4/data.html and choose a tissue type.
Download the R file and you will see it contains a list of Sample IDs for that tissue. This is how the files should be generated.