Read normalized HDF
Create a new HDF that is based on the original HDF, but where the expression values have been normalized. The new files does NOT contain any metadata, so you will still need to read the original HDF for metadata.
Normalization is done with an R script (normalize_expression_data_hpc.R), and it can be quite memory-intensive, so you may want to run in an HPC environment, or on a cloud VM with lots of RAM (~160-200GB).

Also: Updated code to work with `double` types instead of `int` since the normalized values are decimal values, not integers.